% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcde.R
\name{summary_tab}
\alias{summary_tab}
\title{A Summary Table Function}
\usage{
summary_tab(xx, ref = NULL)
}
\arguments{
\item{xx}{Data of which a summary table is needed.}

\item{ref}{A parent data, of which xx is a subset; useful when do stratified analyses}
}
\description{
This function allows you to generate a summary table of your data.
}
\examples{
summary_tab()
}
\keyword{Summary}
\keyword{Table}
